refactor: establish internal/cli, move DuckDB tuning helpers (#529 slice 1)
First slice of #529 (extract the command layer to a shared internal/cli so the
planned bintrail-pg can register the same source-agnostic read commands).

Creates internal/cli and moves the self-contained DuckDB resource-tuning
helpers (shared by query/recover/reconstruct, #510/#511) into it:
AddDuckDBTuningFlags / DuckDBTuningFromFlags / TunedArchiveFetcher (exported)
+ validateDuckDBMemoryLimit + the tests. Repoints query/recover/reconstruct.

Pure no-op refactor — behavior unchanged. Foundation for the follow-up slices
(env-binding helpers, then the source-agnostic commands via cli.AddReadCommands,
then shim). Part of #529.
